agents/safety-plugins/pyproject.toml
[tool.poetry]
name = "safety"
version = "0.1.0"
license = "Apache License 2.0"
description = "Safety Plugins for ADK"
authors = ["sammalik-lgtm <sammalik@google.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
absl-py = "^2.3.1"
google-adk = "^1.7.0"
google-cloud-modelarmor = "^0.2.8"
google-genai = "^1.5.0"
pdfplumber = "^0.11.5"
pydantic = "^2.11.7"
pydantic_core = "^2.33.2"
pydantic-settings = "^2.10.0"
typing_extensions = "^4.14.0"


google-cloud-aiplatform = { extras = [
  "adk",
  "agent-engines",
], version = "^1.93.0" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"